Move controller test utility to separate package to avoid dependency cycle
